[Ubuntu] skip systemctl/journalctl invocation in container environment

+++ images/ubuntu/scripts/build/configure-apt.sh
@@ -4,13 +4,15 @@
 ##  Desc:  Configure apt, install jq and apt-fast packages.
 ################################################################################
 
-# Stop and disable apt-daily upgrade services;
-systemctl stop apt-daily.timer
-systemctl disable apt-daily.timer
-systemctl disable apt-daily.service
-systemctl stop apt-daily-upgrade.timer
-systemctl disable apt-daily-upgrade.timer
-systemctl disable apt-daily-upgrade.service
+if [[ ! -f /run/systemd/container ]]; then
+  # Stop and disable apt-daily upgrade services;
+  systemctl stop apt-daily.timer
+  systemctl disable apt-daily.timer
+  systemctl disable apt-daily.service
+  systemctl stop apt-daily-upgrade.timer
+  systemctl disable apt-daily-upgrade.timer
+  systemctl disable apt-daily-upgrade.service
+fi
 
 # Enable retry logic for apt up to 10 times
 echo "APT::Acquire::Retries \"10\";" > /etc/apt/apt.conf.d/80-retries
